backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.1
SQLAlchemy==2.0.32
alembic==1.13.2
psycopg[binary]==3.2.1
marshmallow==3.21.3
marshmallow-sqlalchemy==1.0.0
python-dotenv==1.0.1
celery==5.4.0
redis==5.0.7
gunicorn==22.0.0
argon2-cffi==23.1.0
email-validator==2.2.0
